+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Unreleased
 
 * Specify a `Keycloak.MIN_TOKEN_VALIDITY_SECONDS` that has a reasonable default for minimum validity and change `Keycloak.updateToken(Runnable)` to use default value rather than `0`.
+* Add `Keycloak.updateTokenAndExecute(Runnable)` override that uses `Keycloak.MIN_TOKEN_VALIDITY_SECONDS` for minimum validity time.
 
 ### [v0.6](https://github.com/realityforge/gwt-keycloak/tree/v0.6) (2020-02-20) · [Full Changelog](https://github.com/realityforge/gwt-keycloak/compare/v0.5...v0.6)